changed to double from float

+++ include/vec3.hpp
@@ -19,20 +19,20 @@ read more here https://shorturl.at/diHV2
 
 class Vec3 {
   public :
-    float e[3]; //e is an arrya of 3 elements of float
+    double e[3]; //e is an arrya of 3 elements of double
 
     Vec3() : e{0,0,0} {} //constructor 1 to create Vec3 with all components to zero
-    Vec3(float e0, float e1, float e2) : e{e0, e1, e2} {} // another constructor is provided that takes three floats as parameters
+    Vec3(double e0, double e1, double e2) : e{e0, e1, e2} {} // another constructor is provided that takes three doubles as parameters
 
-    float x() const {
+    double x() const {
          return e[0]; 
     }
 
-    float y() const {
+    double y() const {
          return e[1]; 
     }
 
-    float z() const {
+    double z() const {
          return e[2];
     }
 
@@ -40,11 +40,11 @@ class Vec3 {
         return Vec3(-e[0], -e[1], -e[2]); 
     }
 
-    float operator[](int i) const { 
+    double operator[](int i) const { 
         return e[i];
     }
 
-    float& operator[](int i) {
+    double& operator[](int i) {
         return e[i];
     }
 
@@ -55,26 +55,26 @@ class Vec3 {
         return *this;
     }
 
-    Vec3& operator*=(float t){
+    Vec3& operator*=(double t){
         e[0] *= t;
         e[1] *= t;
         e[2] *= t;
         return *this;
     }
 
-    Vec3& operator/=(float t) {
+    Vec3& operator/=(double t) {
         return *this *= 1/t;
     }
 
-    float length() const{
+    double length() const{
         return std::sqrt(length_squared());
     }
 
-    float length_squared() const {
+    double length_squared() const {
         return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
     }
 
-    float magnitude() const {
+    double magnitude() const {
         /*Makes more sense to call it magnitude*/
         return std::sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
     }
@@ -120,15 +120,15 @@ inline Vec3 operator*(const Vec3 &u, const Vec3 &v) {
     return Vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-inline Vec3 operator*(float t, const Vec3 &v) {
+inline Vec3 operator*(double t, const Vec3 &v) {
     return Vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 }
 
-inline Vec3 operator/(Vec3 v, float t){
+inline Vec3 operator/(Vec3 v, double t){
     return (1/t) * v;
 }
 
-inline float dot(const Vec3 &u, const Vec3 &v){
+inline double dot(const Vec3 &u, const Vec3 &v){
     return u.e[0] * v.e[0] + u.e[1] * v.e[1] + u.e[2] * v.e[2]; 
 }
